frontend/requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.9.3
numpy==1.26.4
packaging==23.2
pandas==2.2.3
Pillow==9.2.0
protobuf==4.25.5
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.25.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
urllib3==2.2.3
validators==0.34.0
wheel==0.44.0
zipp==3.20.2
